Transport company update
I also added a transport company, currently one, that has a set of transporters (trucks). Based on the state of the containers and the trucks, actions are taken by containers and trucks. Containers can request transport, and a transporter will pickup the container and move towards the container's shipment. The location of both container and transporter are only updated when a goal location (like shipment pickup location) is reached. Thus still very simplistic.